chore: priority value-object refactor + reporting field fixes + menu view-permission-prefix
Not tied to any CA triage ticket — three standalone efforts:

1. Priority PascalCase value object (Request + Appraisal aggregates):
   strict FromString on write / lenient FromDatabase on read, EF HasConversion,
   case-insensitive validators, casing-normalization migrations + parameter reseed.
2. Reporting section-loader field-source fixes (Land/Condo/Income) and
   section-land.html unconditional rendering.
3. Menu viewPermissionPrefix feature (endpoint + reorder handler propagation).
   NOTE: this is NOT CA-396 (Cancel button) — that fix remains unimplemented.

+++ Modules/Reporting/Reporting/Application/Providers/Sections/IncomeSectionLoader.cs
@@ -160,8 +160,8 @@ internal static class IncomeSectionLoader
             var hbuValue = 0m;
             if (!row.IsHighestBestUsed) // C1: HBU land value added when IsHighestBestUsed = false
             {
-                var totalWa = (row.HighestBestUsed_AreaRai ?? 0m) * 400m
-                            + (row.HighestBestUsed_AreaNgan ?? 0m) * 100m
+                var totalWa = (row.HighestBestUsed_AreaRai ?? 0) * 400m
+                            + (row.HighestBestUsed_AreaNgan ?? 0) * 100m
                             + (row.HighestBestUsed_AreaWa ?? 0m);
                 hbuValue = totalWa * (row.HighestBestUsed_PricePerSqWa ?? 0m);
             }
@@ -186,6 +186,7 @@ internal static class IncomeSectionLoader
             IsHighestBestUsed    = row.IsHighestBestUsed,
             HbuAreaRai           = row.HighestBestUsed_AreaRai,
             HbuAreaNgan          = row.HighestBestUsed_AreaNgan,
+            // (int? widens implicitly to decimal? on assignment)
             HbuAreaWa            = row.HighestBestUsed_AreaWa,
             HbuPricePerSqWa      = row.HighestBestUsed_PricePerSqWa,
             YearRows             = yearRows,
@@ -232,8 +233,8 @@ internal static class IncomeSectionLoader
         decimal? FinalValueAdjust,
         decimal? AppraisalPriceRounded,
         bool     IsHighestBestUsed,
-        decimal? HighestBestUsed_AreaRai,
-        decimal? HighestBestUsed_AreaNgan,
+        int?     HighestBestUsed_AreaRai,
+        int?     HighestBestUsed_AreaNgan,
         decimal? HighestBestUsed_AreaWa,
         decimal? HighestBestUsed_PricePerSqWa,
         string?  Summary_GrossRevenueJson,
